2026.02.06 - DeliveryRepositoryTest 테스트 통과 - PaymentRepositoryTest 테스트 통과 - RefundRepositoryTest 테스트 통과

+++ src/main/java/com/example/spring/entity/Payment.java
@@ -34,7 +34,7 @@ public class Payment {
     private PaymentStatus status = PaymentStatus.PENDING;
 
     @Column(nullable = false)
-    private int amount;
+    private BigDecimal amount;
 
     // 결제 일시
     private LocalDateTime paymentDate;
